// src/helpup/db/types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.5";
  };
  public: {
    Tables: {
// ...
      vagas: {
        Row: {
          atividades: string | null;
          capa: string | null;
          categoria: string | null;
          created_at: string;
          descricao: string | null;
          habilidades: string | null;
          id: number;
          local_tempo: string | null;
          ong: string;
          publico: string | null;
          titulo: string;
        };
        Insert: {
          atividades?: string | null;
          capa?: string | null;
          categoria?: string | null;
          created_at?: string;
          descricao?: string | null;
          habilidades?: string | null;
          id?: number;
          local_tempo?: string | null;
          ong: string;
          publico?: string | null;
          titulo: string;
        };
        Update: {
          atividades?: string | null;
          capa?: string | null;
          categoria?: string | null;
          created_at?: string;
          descricao?: string | null;
          habilidades?: string | null;
          id?: number;
          local_tempo?: string | null;
          ong?: string;
          publico?: string | null;
          titulo?: string;
        };
        Relationships: [
          {
            foreignKeyName: "vagas_ong_fkey";
            columns: ["ong"];
            isOneToOne: false;
            referencedRelation: "perfil";
            referencedColumns: ["id"];
          }
        ];
      };
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      [_ in never]: never;
    };
    Enums: {
      "Plan Type": "free" | "premium";
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
};

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">;

type DefaultSchema = DatabaseWithoutInternals[Extract<
  keyof Database,
  "public"
>];
// ...
export type TablesInsert<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema["Tables"]
    | { schema: keyof DatabaseWithoutInternals },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
    : never = never
> = DefaultSchemaTableNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals;
}
  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Insert: infer I;
    }
    ? I
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
  ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
      Insert: infer I;
    }
    ? I
    : never
  : never;
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals;
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
  ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
  : never;
// ...
export type PlanType = Database["public"]["Enums"]["Plan Type"];